Refactoring TileGroup class and adding method to return the existing values that the TileGroup holds.
